Rename to initFirefoxNoHead() Rename pkg to typeregistry
